Completed README
With this, I plan to put this project to rest for a bit (I might revisit it in the summer). My other Maverick team members did not contribute as much as I would have liked (as in very little at all) due to time constraints and have decided to drop this project. As I have put in several hours into this project, I have published it here for others.