Final update for short-term
My team members unfortunately due to lack of time did not contribute much if at all to the end goal of this project (had work, lack of experience, etc) and quit. I plan to halt this project for a while (might revisit it during the summer) and have now posted it on Github with public permissions so that others may benefit.